scripts/requirements.txt
# Contract Auditor Agent - 依赖包

# PDF 处理
pdfplumber>=0.10.0
pdf2image>=1.16.0

# Word 处理
python-docx>=1.0.0

# 图片处理
Pillow>=10.0.0

# OCR 引擎
easyocr>=1.7.0
pytesseract>=0.3.10

# 注意：pytesseract 需要系统安装 Tesseract-OCR
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
# macOS: brew install tesseract
# Linux: apt-get install tesseract-ocr
